Created a default ClientService handler implementation.
Making new ClientService definitions easier.
