+++ client-misk-hibernate/src/test/kotlin/app/cash/backfila/client/misk/ClientMiskTestingModule.kt
@@ -1,8 +1,8 @@
 package app.cash.backfila.client.misk
 
 import app.cash.backfila.client.BackfilaClientConfig
-import app.cash.backfila.client.misk.client.BackfilaClientLoggingSetupProvider
-import app.cash.backfila.client.misk.client.BackfilaClientNoLoggingSetupProvider
+import app.cash.backfila.client.BackfilaClientLoggingSetupProvider
+import app.cash.backfila.client.BackfilaClientNoLoggingSetupProvider
 import app.cash.backfila.embedded.EmbeddedBackfilaModule
 import app.cash.backfila.client.misk.hibernate.ChickenToBeefBackfill
 import app.cash.backfila.client.misk.hibernate.HibernateBackfillModule.Companion.create
