+++ client-misk-jooq/src/test/kotlin/app/cash/backfila/client/misk/config/ClientJooqTestingModule.kt
@@ -2,8 +2,8 @@ package app.cash.backfila.client.misk.config
 
 import app.cash.backfila.client.misk.BackfillModule
 import app.cash.backfila.client.BackfilaClientConfig
-import app.cash.backfila.client.misk.client.BackfilaClientLoggingSetupProvider
-import app.cash.backfila.client.misk.client.BackfilaClientNoLoggingSetupProvider
+import app.cash.backfila.client.BackfilaClientLoggingSetupProvider
+import app.cash.backfila.client.BackfilaClientNoLoggingSetupProvider
 import app.cash.backfila.embedded.EmbeddedBackfilaModule
 import app.cash.backfila.client.misk.jooq.JooqBackfillModule
 import misk.MiskTestingServiceModule
